+++ CHANGELOG.md
@@ -45,6 +45,7 @@
     * **Fancy URLs** renamed to **Pretty URLs**
     * Signature of function `find_url()` changed. Second parameter control absolute or relative URL. Function can build query string from array passed as third parameter
     * Signature of functon `get_page_content()` changed. Added parameter to control replacing page content with page component output
+    * Save user name and date of last saving Components
 
 1. [](#bugfix)
     * `index` page now processed with **Prerry URLs** as other pages. To preserve compatibility user can set custom permalink structure on `index` page to `/`